// pos-frontend/components/buttons/enums.ts
export enum ButtonVariantEnum {
  Primary = 'Primary',
  PrimaryGradient = 'PrimaryGradient',
  PrimaryOutline = 'PrimaryOutline',
  SecondaryOutline = 'SecondaryOutline',
  SecondaryNoline = 'SecondaryNoline',
  White = 'White',
  WhiteOutline = 'WhiteOutline',
}

export enum ButtonSizeEnum {
  sm = 'sm',
  md = 'md',
  lg = 'lg',
  xl = 'xl',
}